squished cuda

+++ lib_cuda/camera.h
@@ -19,15 +19,15 @@ class camera
 {
 public:
 	__host__ __device__ camera() {}
-    __device__ camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect, float aperture, float focus_dist, float t0, float t1)  //vfov is top to bottom degrees
+    __device__ camera(vec3 lookfrom, vec3 lookat, vec3 vup, double vfov, double aspect, double aperture, double focus_dist, double t0, double t1)  //vfov is top to bottom degrees
     {
         time0 = t0;
         time1 = t1;
         lens_radius = aperture / 2;
 
-        float theta = vfov * M_PI / 180;
-        float half_height = tan(theta / 2);
-        float half_width = aspect * half_height;
+        double theta = vfov * M_PI / 180;
+        double half_height = tan(theta / 2);
+        double half_width = aspect * half_height;
         origin = lookfrom;
 
         w = unit_vector(lookfrom - lookat);
@@ -38,11 +38,11 @@ public:
         horizontal = 2*half_width*focus_dist*u;
         vertical = 2*half_height*focus_dist*v;
     }
-    __device__ ray get_ray(curandState *devStates, const int id, float s, float t)
+    __device__ ray get_ray(curandState *devStates, const int id, double s, double t)
     {
         vec3 rd = lens_radius*random_in_unit_disk(devStates, id);
         vec3 offset = u*rd.x() + v*rd.y();
-        float time = time0 + (curand_uniform_double(&devStates[id]) - DELTA) * (time1 - time0);
+        double time = time0 + (curand_uniform_double(&devStates[id]) - DELTA) * (time1 - time0);
         return ray(origin + offset, lower_left_corner + s*horizontal + t*vertical - origin - offset, time);
     }
     vec3 origin;
@@ -50,8 +50,8 @@ public:
     vec3 horizontal;
     vec3 vertical;
     vec3 u, v, w;
-    float time0, time1;
-    float lens_radius;
+    double time0, time1;
+    double lens_radius;
 };
 
 //__device__ camera cam;
