+++ lib_cuda/hitable.h
@@ -7,7 +7,7 @@ class material;
 
 struct hit_record
 {
-    float t;
+    double t;
     vec3 p;
     vec3 normal;
     material *mat_ptr;
@@ -17,7 +17,7 @@ struct hit_record
 class hitable
 {
 public:
-    __device__ virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+    __device__ virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
 };
 
 __device__ hitable *world;
